chore: Simplify rust bridge
