--- apps/hubble/src/addon/addon.ts
@@ -1,22 +1,0 @@
-// This is a bridge between the Rust code and the TS code.
-// We import the Rust code as a NodeJS module, and then export it as a JS function.
-// Note that we need to use the `createRequire` function to import the module, since it
-// is binary code. If we used `import` instead, it would be interpreted as a JS module, and
-// we would get an error becaues it would try to parse it as JS
-import { createRequire } from "module";
-const require = createRequire(import.meta.url);
-const lib = require("./index.node");
-
-export function bridgeEd25519Verify(signature: Uint8Array, hash: Uint8Array, signer: Uint8Array) {
-  const sigBuf = Buffer.from(signature);
-  const hashBuf = Buffer.from(hash);
-  const signerBuf = Buffer.from(signer);
-
-  return lib.ed25519_verify(sigBuf, hashBuf, signerBuf) === 1;
-}
-
-export function bridgeBlake3Hash20(data: Uint8Array) {
-  const dataBuf = Buffer.from(data);
-
-  return new Uint8Array(lib.blake3_20(dataBuf));
-}
